Reorganize examples
